+++ app/src/main/java/com/example/cognizance/domain/models/MovieDetails.kt
@@ -1,14 +1,14 @@
 package com.example.cognizance.domain.models
 
 data class MovieDetails(
-    val id: Long,
+    val id: Int,
     val title: String,
     val backdropPath: String,
     val overview: String,
     val popularity: Double,
     val tagline: String,
     val status: String,
-    val voteAverage: Double,
+    val voteAverage: Float,
     val voteCount: Long,
     val runtime: Long,
     val revenue: Long,
